+++ pkg/types/scheduler.go
@@ -128,9 +128,9 @@ const (
 	ContainerDurationEmissionInterval      time.Duration = 5 * time.Second
 	ContainerResourceUsageEmissionInterval time.Duration = 3 * time.Second
 )
-const ContainerStateTtlSWhilePending float64 = 600
-const ContainerStateTtlS float64 = 120
-const WorkspaceQuotaTtlS float64 = 600
+const ContainerStateTtlSWhilePending int64 = 600
+const ContainerStateTtlS int64 = 120
+const WorkspaceQuotaTtlS int64 = 600
 
 type ErrContainerStateNotFound struct {
 	ContainerId string
